apollo_l1_provider: move records into dedicated file
Only a (chubby) MOVE, no logic changes whatsoever.
This effectively prevents the transaction_manager from directly
manipulating records not through `with_record`, which conserves
invariants and coupling between the records and index.
